Cleaning up startup
When loading the configuration we had been dumping the entire switch
configuration to the log; This was not secure as it included
passwords.

We now add the vlans key to the network model when loading an empty
json object; This is just to make the expected data model more
explicit.